Model 1 and DQC done
Implemented data quality check functions for all models and applied this to model 1. Not yet tested on real data
