+++ data_quality_check.py
@@ -1,0 +1,48 @@
+import os
+import re
+import numpy as np
+import pandas as pd
+import settings
+
+def get_ppt(filename: str):
+   return re.search(r'\d+', filename)[0]
+
+def get_tp(filename:str):
+    return re.search(r'\d+a', filename)[0]
+
+def get_best_cams(checks:list):
+    # checks the quality of each video and returns a pd.Series of the best videos 
+    # from each session
+    # 
+    # checks each csv in frames_checks folder
+    # 
+    # checks parameter is a nested list of groups of keypoints to check
+    # e.g. [['LShoulder', 'LElbow'], ['RShoulder', 'RElbow']] checks
+    # the number of frames where LShoulder and LElbow are both present
+    # and the same for the right side. The mean of all checks is taken
+    # to give a quality score. Passing a standard list will calculate the mean 
+    # proportion of good frames across keypoints (no dependency on both keypoints 
+    # being present).
+
+    camera_scores = []
+    for frame_check in os.listdir(settings.FRAME_CHECKS):
+        data_quality = pd.read_csv(os.path.join(settings.FRAME_CHECKS, frame_check), index_col=0)
+        results = []
+        for check in checks:
+            results.append(data_quality.loc[:, check].all(axis=1).mean())
+        quality_score = np.mean(np.array(results))
+        ppt = get_ppt(frame_check) # get participant from filename
+        tp = get_tp(frame_check) # get timepoint from filename
+        camera_scores.append([os.path.splitext(frame_check)[0], ppt, tp, quality_score])
+
+    camera_scores = pd.DataFrame(camera_scores, columns=['Filename', 'ppt', 'tp', 'QualityScore'])
+    camera_scores.to_csv(settings.ANALYSIS_FOLDER / "model1_camera_scores.csv")
+    print(camera_scores)
+
+    best_cams_idx = list(camera_scores.groupby(['ppt', 'tp'])['QualityScore'].idxmax())
+    best_cams = camera_scores.iloc[best_cams_idx]
+    best_cams.to_csv(str(settings.BEST_CAMERAS))
+
+    return best_cams.Filename
+
+
